generator/requirements.txt
google-cloud-pubsub>=2.18.0
